Move firebaseConfig.js inside /src directory
